#122 [feat] : PC 버전 사이드 바 적용을 위한 분리
